requirements.txt
Flask==2.3.2
pytesseract==0.3.10
Pillow==10.0.0
gunicorn==21.2.0 # Optional: for production deployment